Added a couple more data access classes to the website.
All page sizing is now on each page to allow for every page to be a different width if needed.  Also allows each page to have more control.

+++ Armis.BusinessModels/QualityModels/Spec/SpecRevModel.cs
@@ -7,7 +7,7 @@ namespace Armis.BusinessModels.QualityModels.Spec
     public class SpecRevModel
     {
         public int SpecId { get; set; }
-        public int InternalRev { get; set; }
+        public short InternalRev { get; set; }
         public string Description { get; set; }
         public string ExternalRev { get; set; }
         public int? SamplePlanId { get; set; }
